+++ lib/src/jp/nyatla/nyartoolkit/core/kpm/Utils.java
@@ -1,8 +1,8 @@
 package jp.nyatla.nyartoolkit.core.kpm;
 
 public class Utils {
-	public static float[] arraysubset(float[] a,int s,int l){
-		float[] r=new float[l];
+	public static double[] arraysubset(double[] a,int s,int l){
+		double[] r=new double[l];
 		System.arraycopy(a,s,r,0,l);
 		return r;
 	}
